+++ myHomework/src/homework26/Tunnel.java
@@ -18,7 +18,7 @@ public class Tunnel extends Stage {
                 System.out.println(c.getName() + " готовится к этапу(ждет): " + description);
                 sp.acquire();
                 System.out.println(c.getName() + " начал этап: " + description);
-                Thread.sleep(length / c.getSpeed() * 1000);
+                Thread.sleep((long)(length / c.getSpeed() * 1000));
                 System.out.println(c.getName() + " закончил этап: " + description);
             } catch (InterruptedException e) {
                 e.printStackTrace();
